rename registry.py -> _registry.py to avoid users accidentally doing from autoregistry import registry
