+++ autoregistry/_registry.py
